# Faster uint16 loading in portable decoder