--- a/internal/lz4block/decode_other.go
+++ b/internal/lz4block/decode_other.go
@@ -2,6 +2,8 @@
 
 package lz4block
 
+import "encoding/binary"
+
 func decodeBlock(dst, src []byte) (ret int) {
 	const hasError = -2
 	defer func() {
@@ -32,7 +34,7 @@ func decodeBlock(dst, src []byte) (ret int) {
 					// if the match length (4..18) fits within the literals, then copy
 					// all 18 bytes, even if not all are part of the literals.
 					mLen += 4
-					if offset := uint(src[si]) | uint(src[si+1])<<8; mLen <= offset {
+					if offset := u16(src[si:]); mLen <= offset {
 						i := di - offset
 						end := i + 18
 						if end > uint(len(dst)) {
@@ -66,7 +68,7 @@ func decodeBlock(dst, src []byte) (ret int) {
 			return hasError
 		}
 
-		offset := uint(src[si]) | uint(src[si+1])<<8
+		offset := u16(src[si:])
 		if offset == 0 {
 			return hasError
 		}
@@ -98,3 +100,5 @@ func decodeBlock(dst, src []byte) (ret int) {
 		di += uint(copy(dst[di:di+mLen], expanded[:mLen]))
 	}
 }
+
+func u16(p []byte) uint { return uint(binary.LittleEndian.Uint16(p)) }